+++ src/game/Object/Player.cpp
@@ -2098,209 +2098,6 @@ void Player::RemoveFromWorld()
 }
 
 /**
- * @brief Awards rage generated from dealing or receiving damage.
- *
- * @param damage The raw damage amount used for rage generation.
- * @param attacker True when the player dealt the damage; false when the player received it.
- */
-void Player::RewardRage(uint32 damage, uint32 weaponSpeedHitFactor, bool attacker)
-{
-    float addRage;
-
-    float rageconversion = float((0.0091107836 * getLevel() * getLevel()) + 3.225598133 * getLevel()) + 4.2652911f;
-
-    if (attacker)
-    {
-        addRage = ((damage / rageconversion * 7.5f + weaponSpeedHitFactor) / 2.0f);
-
-        // talent who gave more rage on attack
-        addRage *= 1.0f + GetTotalAuraModifier(SPELL_AURA_MOD_RAGE_FROM_DAMAGE_DEALT) / 100.0f;
-    }
-    else
-    {
-        addRage = damage / rageconversion * 2.5f;
-
-        // Berserker Rage effect
-        if (HasAura(18499, EFFECT_INDEX_0))
-        {
-            addRage *= 1.3f;
-        }
-    }
-
-    addRage *= sWorld.getConfig(CONFIG_FLOAT_RATE_POWER_RAGE_INCOME);
-
-    ModifyPower(POWER_RAGE, uint32(addRage * 10));
-}
-
-/**
- * @brief Regenerates the player's health and power resources for the current tick.
- */
-void Player::RegenerateAll()
-{
-    if (m_regenTimer != 0)
-    {
-        return;
-    }
-
-    // Not in combat or they have regeneration
-    if (!IsInCombat() || HasAuraType(SPELL_AURA_MOD_REGEN_DURING_COMBAT) ||
-        HasAuraType(SPELL_AURA_MOD_HEALTH_REGEN_IN_COMBAT) || IsPolymorphed())
-    {
-        RegenerateHealth();
-        if (!IsInCombat() && !HasAuraType(SPELL_AURA_INTERRUPT_REGEN))
-        {
-            Regenerate(POWER_RAGE);
-        }
-    }
-
-    Regenerate(POWER_ENERGY);
-
-    Regenerate(POWER_MANA);
-
-    m_regenTimer = REGEN_TIME_FULL;
-}
-
-/**
- * @brief Regenerates or decays a specific player power type.
- *
- * @param power The power type to update.
- */
-void Player::Regenerate(Powers power)
-{
-    uint32 curValue = GetPower(power);
-    uint32 maxValue = GetMaxPower(power);
-
-    float addvalue = 0.0f;
-
-    switch (power)
-    {
-        case POWER_MANA:
-        {
-            bool recentCast = IsUnderLastManaUseEffect();
-            float ManaIncreaseRate = sWorld.getConfig(CONFIG_FLOAT_RATE_POWER_MANA);
-            if (recentCast)
-            {
-                // Mangos Updates Mana in intervals of 2s, which is correct
-                addvalue = GetFloatValue(PLAYER_FIELD_MOD_MANA_REGEN_INTERRUPT) *  ManaIncreaseRate * 2.00f;
-            }
-            else
-            {
-                addvalue = GetFloatValue(PLAYER_FIELD_MOD_MANA_REGEN) * ManaIncreaseRate * 2.00f;
-            }
-        }   break;
-        case POWER_RAGE:                                    // Regenerate rage
-        {
-            float RageDecreaseRate = sWorld.getConfig(CONFIG_FLOAT_RATE_POWER_RAGE_LOSS);
-            addvalue = 20 * RageDecreaseRate;               // 2 rage by tick (= 2 seconds => 1 rage/sec)
-        }   break;
-        case POWER_ENERGY:                                  // Regenerate energy (rogue)
-        {
-            float EnergyRate = sWorld.getConfig(CONFIG_FLOAT_RATE_POWER_ENERGY);
-            addvalue = 20 * EnergyRate;
-            break;
-        }
-        case POWER_FOCUS:
-        case POWER_HAPPINESS:
-        case POWER_HEALTH:
-            break;
-        default:
-            break;
-    }
-
-    // Mana regen calculated in Player::UpdateManaRegen()
-    // Exist only for POWER_MANA, POWER_ENERGY, POWER_FOCUS auras
-    if (power != POWER_MANA)
-    {
-        AuraList const& ModPowerRegenPCTAuras = GetAurasByType(SPELL_AURA_MOD_POWER_REGEN_PERCENT);
-        for (AuraList::const_iterator i = ModPowerRegenPCTAuras.begin(); i != ModPowerRegenPCTAuras.end(); ++i)
-        {
-            if ((*i)->GetModifier()->m_miscvalue == int32(power))
-            {
-                addvalue *= ((*i)->GetModifier()->m_amount + 100) / 100.0f;
-            }
-        }
-    }
-
-    if (power != POWER_RAGE)
-    {
-        curValue += uint32(addvalue);
-        if (curValue > maxValue)
-        {
-            curValue = maxValue;
-        }
-    }
-    else
-    {
-        if (curValue <= uint32(addvalue))
-        {
-            curValue = 0;
-        }
-        else
-        {
-            curValue -= uint32(addvalue);
-        }
-    }
-
-    SetPower(power, curValue);
-}
-
-/**
- * @brief Regenerates the player's health based on state, auras, and rates.
- */
-void Player::RegenerateHealth()
-{
-    uint32 curValue = GetHealth();
-    uint32 maxValue = GetMaxHealth();
-
-    if (curValue >= maxValue)
-    {
-        return;
-    }
-
-    float HealthIncreaseRate = sWorld.getConfig(CONFIG_FLOAT_RATE_HEALTH);
-
-    float addvalue = 0.0f;
-
-    // polymorphed case
-    if (IsPolymorphed())
-    {
-        addvalue = (float)GetMaxHealth() / 3;
-    }
-    // normal regen case (maybe partly in combat case)
-    else if (!IsInCombat() || HasAuraType(SPELL_AURA_MOD_REGEN_DURING_COMBAT))
-    {
-        addvalue = OCTRegenHPPerSpirit() * HealthIncreaseRate;
-        if (!IsInCombat())
-        {
-            AuraList const& mModHealthRegenPct = GetAurasByType(SPELL_AURA_MOD_HEALTH_REGEN_PERCENT);
-            for (AuraList::const_iterator i = mModHealthRegenPct.begin(); i != mModHealthRegenPct.end(); ++i)
-            {
-                addvalue *= (100.0f + (*i)->GetModifier()->m_amount) / 100.0f;
-            }
-        }
-        else if (HasAuraType(SPELL_AURA_MOD_REGEN_DURING_COMBAT))
-        {
-            addvalue *= GetTotalAuraModifier(SPELL_AURA_MOD_REGEN_DURING_COMBAT) / 100.0f;
-        }
-
-        if (!IsStandState())
-        {
-            addvalue *= 1.5;
-        }
-    }
-
-    // always regeneration bonus (including combat)
-    addvalue += GetTotalAuraModifier(SPELL_AURA_MOD_HEALTH_REGEN_IN_COMBAT);
-
-    if (addvalue < 0)
-    {
-        addvalue = 0;
-    }
-
-    ModifyHealth(int32(addvalue));
-}
-
-/**
  * @brief Gets an NPC the player can currently interact with.
  *
  * @param guid The target creature GUID.
@@ -5427,16 +5224,6 @@ bool Player::CanSpeak() const
     return  GetSession()->m_muteTime <= time(NULL);
 }
 
-/*********************************************************/
-/***              LOW LEVEL FUNCTIONS:Notifiers        ***/
-/*********************************************************/
-
-void Player::SendAttackSwingNotInRange()
-{
-    WorldPacket data(SMSG_ATTACKSWING_NOTINRANGE, 0);
-    GetSession()->SendPacket(&data);
-}
-
 /**
  * @brief Replaces a tokenized uint32 field value in a serialized array.
  *
@@ -5463,51 +5250,6 @@ void Player::SetUInt32ValueInArray(Tokens& tokens, uint16 index, uint32 value)
 void Player::SendAttackSwingNotStanding()
 {
     WorldPacket data(SMSG_ATTACKSWING_NOTSTANDING, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the error packet for attempting to attack a dead target.
- */
-void Player::SendAttackSwingDeadTarget()
-{
-    WorldPacket data(SMSG_ATTACKSWING_DEADTARGET, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the error packet for a general inability to attack the target.
- */
-void Player::SendAttackSwingCantAttack()
-{
-    WorldPacket data(SMSG_ATTACKSWING_CANT_ATTACK, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the packet that cancels the player's current attack.
- */
-void Player::SendAttackSwingCancelAttack()
-{
-    WorldPacket data(SMSG_CANCEL_COMBAT, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the error packet for attempting to attack while facing the wrong direction.
- */
-void Player::SendAttackSwingBadFacingAttack()
-{
-    WorldPacket data(SMSG_ATTACKSWING_BADFACING, 0);
-    GetSession()->SendPacket(&data);
-}
-
-/**
- * @brief Sends the packet that cancels auto-repeat attacks for the client.
- */
-void Player::SendAutoRepeatCancel()
-{
-    WorldPacket data(SMSG_CANCEL_AUTO_REPEAT, 0);
     GetSession()->SendPacket(&data);
 }
